requirements.txt
streamlit
openai
pandas
PyMuPDF
python-dotenv
requests
pdfplumber
PyPDF2
sqlalchemy
tqdm